chore: stop using merge groups
Unfortunately, the squash-before-merge behavior seems to be broken or non-existent.

The goal of using merge groups was to avoid changelog spam, where a commit would be generated every time a beta was released. Instead, the changelog update would be amended to the last PR in the merge group.

The next best option I could think of is to limit beta releases to once per day (currently set to 5 AM UTC, a.k.a. 1 AM eastern time).