chore: add `pnpm test-single` command
Use it to run tests for a single function (or a group of functions that all start with the same letters).

    pnpm test-single shake

The benefit is that test coverage will only be reported for that specific function. If you did `pnpm test shake`, test coverage will be polluted with false negatives (because Vitest will think all the other functions don‘t have tests, since they didn‘t run).