ci: pull in PR changes for pull_request_target triggers
In order to access repository secrets in a workflow that is triggered by a pull request, we need to use the pull_request_target event. But that means the PR is never checked out, so we have to do it manually. We only check out the files we need, to prevent scripts from being maliciously overwritten, thereby leaking our repository secrets.