ci: fix bench-pr workflow
Since we are using the "pull_request_target" event to trigger the workflow, the PR files are checked out using `git checkout` with a file filter. That means the PR changes are staged in the working tree, so we should use `git diff --staged` (instead of providing a base ref) to determine which files are changed/added by the PR.

This commit also includes some cosmetic changes.