chore: switch from esnext to es2017 in tsconfig
The goal here is to prevent use of JS APIs that are not supported by older browsers that we want to support out-of-the-box without legacy transforms. We were using eslint-plugin-compat, but it doesn‘t check method use, so I‘ve decided to remove it. This also has a bonus of removing the dependency on eslint, so now we only have one linter in the project (Biome).