chore: make merge function work faster with large arrays
Changed the `array.indexOf` to Map to get rid of the nested loop and
make the function work faster with large arrays. The function performs
a slight worse for the small arrays, but the difference is negligible.
With a big arrays it's significantly faster. The benchmark shows that
for arrays with 10000 elements the new implementation is 10 times
faster.